Dashboard

+++ app/Http/Controllers/CategoryController.php
@@ -2,11 +2,14 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Account;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
-
 use App\Models\Category;
 use App\Models\Idea;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Validation\Validator;
 
 class CategoryController extends Controller
 {
@@ -16,11 +19,13 @@ class CategoryController extends Controller
     }
 
     public function store(Request $request){
-
+        $request->validate([
+            'category_name' => 'required|max:255'
+        ]);
         $cate = Category::create([
             'category_name' => $request->category_name,
-            'first_closure_date' => Carbon::now(),
-            'second_closure_date' => Carbon::now()->addDays(14),
+            'first_closure_date' => $request->category_date,
+            'second_closure_date' => Carbon::createFromFormat('Y-m-d', $request->category_date)->addDays(14),
         ]);
         if(!$cate){
             return redirect()->back()->with('error','Create Category Not Success');
@@ -30,10 +35,15 @@ class CategoryController extends Controller
     }
 
     public function update(Request $request, $id){
-
+//        if (! Gate::allows('edit-cate', $account)) {
+//            return redirect()->back()->with('error','You can not edit category');
+//        }
         $cate = Category::find($id);
 
         if($cate){
+            $request->validate([
+                'category_name' => 'required'
+            ]);
             $update_cate = $cate->update([
                 'category_name' => $request->input('category_name'),
             ]);
@@ -44,8 +54,14 @@ class CategoryController extends Controller
         }
     }
 
-    public function destroy(Request $request, $id){
+    public function destroy(Request $request, $id, Category $category){
+//        if (! Gate::allows('delete-cate', $category)) {
+//            return redirect()->back()->with('error','You can not delete category');
+//        }
         $cate = Category::find($id);
+        if (count($cate->idea) != 0){
+            return redirect()->back()->with('error','Delete Category Not Success');
+        }
         if($cate){
             $delete_cate = $cate->delete();
